chore: move captureWithTabSourceId() to a shared helper
